+++ src/hooks/useTodoHandlers.js
@@ -1,0 +1,18 @@
+import {
+  handleAddTodo,
+  handleToggleComplete,
+  handleDeleteTodo,
+  handleSetFilter,
+} from "../todoHandlers";
+
+const useTodoHandlers = (todos, setTodos, setFilter) => {
+  return {
+    handleAddTodo: (todo) => handleAddTodo(todos, setTodos, todo),
+    handleToggleComplete: (index) =>
+      handleToggleComplete(todos, setTodos, index),
+    handleDeleteTodo: (index) => handleDeleteTodo(todos, setTodos, index),
+    handleSetFilter: (filter) => handleSetFilter(setFilter, filter),
+  };
+};
+
+export default useTodoHandlers;
