+++ src/hooks/useTodos.js
@@ -1,28 +1,18 @@
-import useTodosState from "./useTodosState";
-import useFilteredTodos from "./useFilteredTodos";
-import {
-  handleAddTodo,
-  handleToggleComplete,
-  handleDeleteTodo,
-  handleSetFilter,
-} from "../todoHandlers";
+import useTodoStateManager from "./useTodoStateManager";
+import useTodoHandlers from "./useTodoHandlers";
 import FILTERS from "../todoUtils/filters";
 
-// 할 일 목록과 관련된 상태와 함수를 제공하는 훅
 const useTodos = () => {
-  const { todos, setTodos, filter, setFilter } = useTodosState();
-  const filteredTodos = useFilteredTodos(todos, filter);
+  const { todos, allTodos, setTodos, filter, setFilter } =
+    useTodoStateManager();
+  const handlers = useTodoHandlers(allTodos, setTodos, setFilter);
 
   return {
-    todos: filteredTodos,
-    allTodos: todos,
-    handleAddTodo: (todo) => handleAddTodo(todos, setTodos, todo),
-    handleToggleComplete: (index) =>
-      handleToggleComplete(todos, setTodos, index),
-    handleDeleteTodo: (index) => handleDeleteTodo(todos, setTodos, index),
-    handleSetFilter: (filter) => handleSetFilter(setFilter, filter),
+    todos,
+    allTodos,
     filter,
     FILTERS,
+    ...handlers,
   };
 };
